requirements.txt
certifi==2023.7.22
elastic-transport==8.4.0
elasticsearch==8.8.2
psycopg2-binary==2.9.7
urllib3==1.26.16
